Btrfs-progs: join DEV_STATS ioctls to one
On June 22th 2012, David Sterba posted a patch to change the ioctls
for the commands to get/reset device stats in the kernel.
Intentionally, this was not backward compatible since the goal was to
get rid of one of the two ioctls. And it was early enough before any
Linux release included the device statistics feature. Only some kernel
RCs include the old ioctl definitions.

This patch here adapts the btrfs-progs side to use the modified ioctl
definitions, also intentionally without backward compatibility, with
the justification, that no Linux kernel release contains the old
definitions.

Signed-off-by: Stefan Behrens <sbehrens@giantdisaster.de>